requirements.txt
google-genai
feedparser
gspread
python-dotenv
lxml
